Fix render error and add return button to books
Rendering the list of books the user checked out
is dependent on the books property of the user
object. I tried to check the content of that
property before the user was fetched, which caused
an error. I fixed it by checking if the books
property exists before checking if there are any
books in it.